LPD-78234 Ignore reverse relationships since they cannot be removed

+++ modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/struts/UpdateStructureStrutsAction.java
@@ -113,8 +113,10 @@ public class UpdateStructureStrutsAction implements StrutsAction {
 					getObjectRelationshipsByObjectDefinitionId2(
 						objectDefinitionId, false)) {
 
-			_objectRelationshipLocalService.deleteObjectRelationship(
-				objectRelationship);
+			if (!objectRelationship.isReverse()) {
+				_objectRelationshipLocalService.deleteObjectRelationship(
+					objectRelationship);
+			}
 		}
 	}
